+++ jvm/selfie-runner-kotest/src/commonMain/kotlin/com/diffplug/selfie/kotest/SnapshotSystemKotest.kt
@@ -29,7 +29,6 @@ import com.diffplug.selfie.guts.InlineWriteTracker
 import com.diffplug.selfie.guts.LiteralValue
 import com.diffplug.selfie.guts.SnapshotFileLayout
 import com.diffplug.selfie.guts.SnapshotSystem
-import com.diffplug.selfie.guts.SourceFile
 import com.diffplug.selfie.guts.ToBeFileWriteTracker
 import com.diffplug.selfie.guts.TypedPath
 import com.diffplug.selfie.guts.WithinTestGC
@@ -88,7 +87,7 @@ internal class SnapshotSystemKotest(settings: SelfieSettingsAPI) : SnapshotSyste
 
     if (mode != Mode.readonly) {
       for (path in commentTracker.pathsWithOnce()) {
-        val source = SourceFile(path.name, layout.fs.fileRead(path))
+        val source = layout.parseSourceFile(path)
         source.removeSelfieOnceComments()
         layout.fs.fileWrite(path, source.asString)
       }